feat: implement stock price adjustment handling and update holdings snapshot calculations

+++ bluesky-parent/bluesky-web/bluesky-web-gate/src/main/java/net/luversof/web/gate/stock/dto/response/HoldingsSnapshotItem.java
@@ -5,7 +5,7 @@ import java.math.BigDecimal;
 public record HoldingsSnapshotItem(
         String name,
         String symbol,
-        long quantity,
+        BigDecimal quantity,
         BigDecimal avgCost,
         BigDecimal priceAtDate,
         BigDecimal value,
